Safari-609.1.11.1/LayoutTests/pointer-lock/lock-lost-on-navigation-expected.txt

Test that pointer lock is lost after navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.testRunner is defined.
PASS internals.pageHasPendingPointerLock() is false
PASS internals.pageHasPointerLock() is false
PASS successfullyParsed is true

TEST COMPLETE
